Double render Fixed
