+++ src/components/ProductsList.jsx
@@ -10,20 +10,20 @@ class ListProducts extends React.Component {
         <section>
           <div className="product-list">
             { products.map((product) => (
-              <Link
-                to={ `/product/${product.id}` }
-                key={ product.id }
-                data-testid="product-detail-link"
-              >
-                <div className="product-card" data-testid="product">
+              <div className="product-card" data-testid="product" key={ product.id }>
+                <Link
+                  to={ `/product/${product.id}` }
+                  data-testid="product-detail-link"
+                >
                   <h4>{ product.title }</h4>
                   <img src={ product.thumbnail } alt={ product.title } />
                   <h3>
                     { 'R$ ' }
                     { product.price }
                   </h3>
-                </div>
-              </Link>)) }
+                </Link>
+              </div>
+            )) }
           </div>
         </section>
       );
