+++ src/pages/ShoppingCart.jsx
@@ -5,10 +5,6 @@ class ShoppingCart extends Component {
   constructor() {
     super();
 
-    if (!localStorage.getItem('cartProducts')) {
-      localStorage.setItem('cartProducts', '[]');
-    }
-
     this.state = {
       products: [],
     };
@@ -18,13 +14,6 @@ class ShoppingCart extends Component {
     this.getLocalstorage();
   }
 
-  saveAtLocalstorage = (objectToSave) => {
-    let localObj = JSON.parse(localStorage.getItem('cartProducts'));
-    localObj = [...localObj, objectToSave];
-
-    localStorage.setItem('cartProducts', JSON.stringify(localObj));
-  }
-
   getLocalstorage = () => {
     this.setState({
       products: JSON.parse(localStorage.getItem('cartProducts')),
@@ -32,47 +21,23 @@ class ShoppingCart extends Component {
   }
 
   render() {
-    const { location: { state } } = this.props;
     const { products } = this.state;
-    if (state) {
-      const { title, price, imgUrl } = state;
-      console.log(title, price, imgUrl);
-      this.saveAtLocalstorage(state);
-      return (
-        <section>
-          <h1>Carrinho de Compras</h1>
-          { products.map((product) => <p key={ product.title }>{product.title}</p>) }
-        </section>
-      );
-    }
-
+    console.log(products);
     return (
       <section>
         <h1>Carrinho de Compras</h1>
-        <p data-testid="shopping-cart-empty-message">Seu carrinho está vazio</p>
+        { products.length
+          ? products.map((product) => <p key={ product.title }>{product.title}</p>)
+          : <p>Seu Carrinho está vazio!</p>}
       </section>
     );
   }
 }
 
-ShoppingCart.defaultProps = {
-  location: {
-    state: {
-      title: '',
-      price: '',
-      imgUrl: '',
-    },
-  },
-};
-
 ShoppingCart.propTypes = {
   location: PropTypes.shape({
-    state: PropTypes.shape({
-      title: PropTypes.string,
-      price: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-      imgUrl: PropTypes.string,
-    }),
-  }),
+    state: PropTypes.shape({}),
+  }).isRequired,
 };
 
 export default ShoppingCart;
